Create Minimal VII: Lambda Functions.py
In the Code tab you will find a code that is missing a single character in order to pass the tests. However, your goal is to submit a function as minimalist as possible. Use the tips in the tips section below.

Write five adder functions:

add_2(x) should return 2 + x.
add_3(x) should return 3 + x.
add_5(x) should return 5 + x.
add_7(x) should return 7 + x.
add_11(x) should return 11 + x.
Tips
Functions that consists only of a return, can be written as oneliner with a lambda function.

For example, the code:

def are_same(a, b):
    return a == b
Can be simplified to:

are_same = lambda a, b: a == b
Bonus
lambda a, b: a == b is technically an anonymous function. However, it can be assigned to the identifier are_same and it can then be called using are_same().