Create Clear the Fog.py
Clear the Fog
Create a function which returns the word in the string, but with all the fog letters removed. However, if the string is clear from fog, return "It's a clear day!".

Examples
clear_fog("sky") ➞ "It's a clear day!"

clear_fog("fogfogFFfoooofftogffreogffesGgfOogfog") ➞ "trees"

clear_fog("fogFogFogffoObirdsanffodthebffoeffoesGGGfOgFog") ➞ "birdsandthebees"
Notes
Hidden words won't include the letters f, o or g.
Hidden words are always in lowercase.